Haproxy driver should respect vip/pool admin state
On vip/pool update when admin_state_up becomes False
haproxy driver should reflect it in the config.
Currently there may be only one vip in the config and
if it is disabled, haproxy process fails to restart with
"[ALERT] 084/045122 (11407) : [haproxy.main()] No enabled
listener found (check the <listen> keywords) ! Exiting.",
and continues running and balancing with old config -
so for this case we need to undeploy loadbalancer.
The patch also moves namespace deletion to delete_pool()
as there is no need to delete/recreate namespace each time
the vip is removed/added or disabled/enabled

Closes-Bug: #1155092
Closes-Bug: #1297142
Change-Id: I11e2bd3185328ba47ba1aaede932e3114263bed8