LBaaS: Unit tests using policy.conf
Instead of adding override_nvalue() in setUp() for tests that need
the policy.json setting for config, created a new base class for
VPN tests, which derives from the neutron BaseTestCase and overrides
the config for policy_file. Applying same changes as done with
141932 for VPNaaS.

The big question here, is whether or not this is better than
sprinkling override_nvalues() calls in the four places, where this
is a problem currently.  The issue is that with LBaaS, there are
some tests that inherit from Neutron classes, other than the
BaseTestCase, so a simple override in BaseTestCase.setUp() won't
suffice.

Change-Id: Ia504df51510d964c5f5baccb0b921af1de406f21
Partially-Implements: blueprint service-split