Persistence lbaas http cookie
when using lbaas haproxy session persistence of type HTTP_COOKIE, the
lbaas haproxy driver creates haproxy config with entries for pool
members, in which the value of 'cookie' argument is placed in session
cookies and used to route requests to correct pool member, ensuring
session persistence, but when a new pool member is created or deleted,
a new cookie value for member is created. This causes old HTTP sessions
to be routed to new pool member, and session is broken.
Fixes the problem by using member id as cookie value.

Change-Id: I0b23dab10c541c67635a7d9ba3a9e0ac0a75ac94
Closes-bug: #1404248