Block subnet gateway IP to be used as LB VIP
This fix raises an IpAddressInUse exception when subnet gateway IP is assigned
to a loadbalancer VIP.

For a neutron network that is connected to a hardware router, its subnet
gateway IP address is not associated with a neutron port. But neutron allows
user to set this gateway IP as a loadbalancer VIP.  When both the router and LB
VIP uses the same IP address, traffic to and from the subnet may not reach its
intended destination.

Change-Id: I97fb0208f4178fad1a3c5c2a80604661817c448a
Closes-Bug: 1391059