Added tox tempest environment
When running tox with tox -etempest, tox will install tempest
from a specific commit that the v2 tests are known to work well
with.  Then the tests will be run. This also puts a change in
pretty_tox.sh to control the concurrency level because the
tempest tests will fail when running in parallel.

Change-Id: I745293322149390f0e709165f34c144f5c439720