Do not assume http expected_codes ordering
This fixes the unit tests[1] that breaks with a randomized
PYTHONHASHSEED (see the bug report).

The test assumed that expected_codes had elements in particular order.
Found with PYTHONHASHSEED=1.

The fix refactors the test case by injecting values using the same
ordering.

[1] neutron_lbaas.tests.unit.services.loadbalancer.drivers.haproxy.\
    test_jinja_cfg

Partial-bug: #1348818

Note: There are several other unrelated unit tests that also break with
a randomized PYTHONHASHSEED, but they are not addressed here. They will
be addressed in separate patches.

Change-Id: I5d761c392bf37a0b3626abc64e0e428f978638df