Use "session_persistence" in data models for uniformity with API attribute
Currently API uses "session_persistence" attribute name for session
persistence while in data model the name is "sessionpersistence".
This naming difference makes things complicated and leads to errors
in case  wrong attribute name usage.

For example, if we try to create a pool with defined session persistence
from_dict method of the Pool data model tries to pop "session_persistence"
from a model_dict [1]. But in reality model_dict contains "sessionpersistence"
key due to the replacement that was done in from_sqlalchemy_model method [2]
according to the db model [3]. This will later lead to errors during
refreshing of a device when pool.sessionpersistence [4] returns a dict
instead of a SessionPersistence data model.

Kept sessionpersistence attribute in Pool data model for backwards
compatibility.

[1] https://github.com/openstack/neutron-lbaas/blob/master/neutron_lbaas/services/loadbalancer/data_models.py#L350
[2] https://github.com/openstack/neutron-lbaas/blob/master/neutron_lbaas/services/loadbalancer/data_models.py#L67
[3] https://github.com/openstack/neutron-lbaas/blob/master/neutron_lbaas/db/loadbalancer/models.py#L148
[4] https://github.com/openstack/neutron-lbaas/blob/master/neutron_lbaas/services/loadbalancer/drivers/haproxy/jinja_cfg.py#L297

Change-Id: Ib5efac45cf19acf0afe88a7ffa79514dd4129ed1
Closes-Bug: #1477348