Add ability for octavia to allocate vip
Adds a configuration option (defaults to False) to allow octavia to be
responsible for allocating the VIP.  If this is the case, then the driver will
then update the neutron-lbaas load balancer database with the vip information.

This satisfies a few use cases where a single port created by neutron lbaas
on behalf of the user is too restrictive for more complex vip allocation needs
that Octavia does and will support.

Closes-Bug: #1463594
Change-Id: Idd52d5d8325f1447aadcdf9c687ce010a6488cf6