Change tls and sni listener API attributes to ref
It was originally thought that neutron lbaas would take a uuid for a tls
container that references a container of secrets.  This has changed due to the
fact that we will now take a URI for barbican containers, which makes the use
of default_tls_container_id and sni_container_ids inappropriate.  This changes
them to be default_tls_container_ref and sni_container_refs.

This patch does not change the db models or data models as that would make this
a much larger patch and due to time would be a lot to merge in by the
Liberty-3.  Since V2 API will probably lose the experimental tag this cycle, it
is important to make this change now while it is experimental.
This just changes the API to accept and return these attributes so
whenever the db and data model changes are done, it will not break API
backwards compatibility.

DocImpact
APIImpact

Change-Id: Iddc5d95fc6822e19e3166f9a6974dc56edd3193e