Add flavor option to loadbalancerv2 creation
Builds on the Neutron services flavor framework in the referenced
Depends-On change id.

This solution uses a flavor of service type LOADBALANCERV2
associated with a flavor profile containing a driver.  If a
flavor_id is passed during loadbalancerv2 creation, the flavor is
used to find the provider for the driver and populate
the existing lbaas provider field.

This allows a user to specify a flavor like 'Gold' with the
operator dynamically controlling which provider is associated
with that flavor for subsequent creates.

The flavor_id is persisted in the DB.  The selected provider
controls subsequent control operations against the created
loadbalancer.

Also ensure everything works gracefully if flavors plugin is not loaded.

Adding api and doc impact tags as the new optional flavor_id parameter
should be described to consumers.

Co-Authored-By: James Arendt <james.arendt@hp.com>
Co-Authored-By: Madhusudhan Kandadai <madhusudhan.kandadai@hpe.com>

ApiImpact
DocImpact
Change-Id: Ic3b459692d092658df93d4c855079311d9bc7ace
Depends-On: I5c22ab655a8e2a2e586c10eae9de9b72db49755f
Partially implements: blueprint neutron-flavor-framework