Various fixes for tempest tests with octavia
The octavia driver is now sending valid key values to octavia on listener and
pool updates.

Tempest tests are now appropriately waiting for the load balancer to become
active on many tests that made a change to the load balancer or its child but
did not wait for the load balancer to become active before making another
change to that load balancer or one of its children.

Change-Id: I0dce23401010b02a469d34e20adc22f020b83dbd