Use admin context in octavia driver, improve logs
When polling for the status of a load balancer in octavia, the context created
by the plugin was being used.  Since the polling takes place in a thread, that
context can have some unexpected behavior.  It is better to create an admin
context in the thread because it definitely exists in that thread.

Also improve debug logging in the octavia driver and the successful and failed
completion helper methods in the mixins.

Change-Id: I6f64c95d89876c9d1d19076a4fe7e3006a0df5e0
Closes-Bug: #1497823