Fix scenario tests for octavia driver
The current scenario tests were not waiting for the load balancer to become
ACTIVE while doing cleanups causing test failures.

Also, this increases a hardcoded timeout value of when a load balancer is
deemed to be stuck.  It was too short for the gate and the nested VMs issue of
boot taking 5-8 minutes.  I've increased it to 10 minutes.

A timeout was also needed when testing connectivity to a VIP because the
default timeout is way too long

Change-Id: I667af547296ac707a6176d5c400ce025366e65ba
(cherry picked from commit cb4af16b532e93c475be626692bc3c7d8741d072)