Remove dependency on neutron for topics
LBaaS uses LOADBALANCER_PLUGIN and LOADBALANCER_AGENT from neutron topics.py.
Copying these to LBaaS project, as not used in neutron, and will remove from
neutron later.

Also, see LOADBALANCER and LOADBALANCERV2. The former is only used in neutron
tests (so copying to LBaaS project). The latter is not used in neutron,
so placing here and it can be removed from neutron later.

The goal here is to chip away at removing dependencies on the neutron
repo. This is just one simple case to clean up some unneeded references.

Placed the literals in existing constants file in LBaaS project.

Change-Id: I62bfd6b34c381427425ca2a656e5782087de3a3b
Implements: blueprint neutron-lib