Add an explicit BRANCH_NAME to tox_install.sh
tools/tox_install.sh exists to deal with installing neutron without
neutron existing as a pypi package.

The way this operates causes problems in the periodic jobs on the stable
branches.  Make it explicit which branch to install in the gate
(zuul-cloner) and via pip.

This is done with a BRANCH_NAME variable to make it trivial to change
once "master" is tagged as stable.

Conflicts:
    tools/tox_install.sh

The conflict is due to I0e8749f49230b28bcd554b302a9277283c428cc6 not
existing in stable/liberty

Change-Id: I0667580c3107d3c1e1888a1b2c2a222124d22fe3
(cherry picked from commit 9557c124085aa0e83b8558b152323b60a0a3e5eb)