Add an explicit BRANCH_NAME to tox_install.sh
tools/tox_install.sh exists to deal with installing neutron without
neutron existing as a pypi package.

The way this operates causes problems in the periodic jobs on the stable
branches.  Make it explicit which branch to install in the gate
(zuul-cloner) and via pip (NEUTRON_PIP_LOCATION).

This is done with a BRANCH_NAME variable to make it trivial to change
once master is tagged as stable.

Change-Id: I0667580c3107d3c1e1888a1b2c2a222124d22fe3