Delete tempest fork
A while ago we copied Tempest networking API tests in to the Neutron
repo, and along came thousands of lines of code of Tempest testing
infrastructure (neutron.tests.tempest). For a while we periodically
refreshed our fork via: tools/copy_api_tests_from_tempest.sh

It's time to remove tempest fork[1] and use tempest_lib when possible or
tempest itself for code not moved to tempest-lib instead of duplicating,
forking tempest as done in Neutron, Manila & Ironic.

This change is based on the similar work done in Neutron tree[2].

[1] neutron_lbaas.tests.tempest.lib
[2] I4f9193dfe26f2d36985cb480a98709ec182a2f7b

Co-Authored-By: Franklin Naval <franklin.naval@gmail.com>

Closes-Bug: #1536852
Change-Id: I516b37d45200cc120a929ecd9011ff8bb4c45d9f