Auto reschedule loadbalancers from dead agents
Similarly to what was done in 'allow_automatic_l3agent_failover'
This patch adds a periodic check to examine the status of all LBaaS
agents with loadbalancers scheduled to them.
If the agent is dead, the loadbalancer will be rescheduled to an alive
agent.

The periodic LBaaS agents status check is invoked by
add_agent_status_check_worker() which was introduced in neutron codebase in
I3a32a95489831f0d862930384309eefdc881d8f6 to allow safe process forking.

Closes-Bug: #1565511

Depends-On: I652ab029b7427c8783e4b2f0443a89ee884bf064
Change-Id: Id8d3218bf1e52722cc10ddcd34e3e734eef90658