LBaaS V1 Log error instead of exception
A test was failing because the check for pool orphans was doing a
LOG.exception when there was no exception ever thrown during a unit
test, so the LOG.exception call was failing.  It doesn't make sense
for this piece of code to LOG.exception because there is no
exception being thrown, its throwing its own exception.

Change-Id: I21c12b639d05076f6c89e5c023a60526e81db5b7
Closes-Bug: #1578771