Skip the check _check_pool_loadbalancer_match in one condition
In function create_listener,_check_pool_loadbalancer_match is used to
make sure default_pool has the same lb_id as given.
When a listener is created with default_pool_id and no lb_id,
lb_id is set to default_pool.loadbalancer.id
So the check _check_pool_loadbalancer_match does repeated.
I want to skip this extra check in the condition that a listener is
created with default_pool_id and no lb_id.

Change-Id: Ie6d3d44496faf9684fc4ddb1e0c952f089d0e249
Closes-Bug: #1588118