Cap tempest to < 12.1.0
Latest tempest release is incompatible with urllib3 1.14 that is
enforced in gate using upper constraints. Tempest uses chunked= argument
to construct a request, the argument that is introduced in urllib3 1.15+
only.

This patch is not a backport from master because it's not needed there.
On master, urllib3 is fresh enough to pass the gate.

Change-Id: I3b48b86ffe75f3c12b4d13e0239530f18744ef69