Fix LBaaS gate
Changes 9e227c5b6c and 546ba377db brought disaster and despair.
The former introduced a new utility method to tempest.lib and
removed the private method _try_delete_resource. A new release
of tempest broke the LBaaS API tests.
The latter fixed migration tests and in the process it shuffled
a few things around that ended up breaking the LBaaS unit tests.

This patch restores sanity, if only for so long.

Change-Id: Ib47aa60053c53f126cf7556bad345b8ac278812c