haproxy ns driver: set global haproxy maxconn
Haproxy uses two separate maxconn values: the maxconn setting in the
global context is used to adjust file descriptor limits and allocate
resources, while the maxconn setting in the frontend context controls
if additional connections are accepted on that frontend.

This patch adds a maxconn to the global context equal to the sum of the
connection limit on each listener.

Change-Id: I950abaed780d7df33f0ab5c682bf583aa581889f
Closes-Bug: 1544861