haproxy ns driver: set global haproxy maxconn
Haproxy uses two separate maxconn values: the maxconn setting in the
global context is used to adjust file descriptor limits and allocate
resources, while the maxconn setting in the frontend context controls if
additional connections are accepted on that frontend.

This patch adds a maxconn to the global context equal to the sum of the
connection limit on each listener. For listeners which no connection
limit specified, the haproxy default global connection limit (2000
connections) is used. This way we provide the same behavior as a default
haproxy configuraiton which no connection limit specified in the case of
a single load balancer.

Change-Id: Icc63771ed15a7f718dab0389c1fb63d628896e0a
Partial-Bug: 1544861