The lb namespace can not be deleted
When delete the last listener, the lb instance was undeployed.
But the lb namespace was not deleted at this time. After that,
when delete the loadbalancer, the lb namespace will not be
deleted for the exists() check always return False.

Closes-Bug: #1495430
(cherry picked from commit 9073940a82d3adf6581643941c059bf1d8bb8615)
Change-Id: Ic3cb25a6bd2bf86f8c30ef62ae323c3bab0c82fc