Use temporary directory for neutron install
Do not install in /tmp/openstack/neutron and leave the git repository
there after the script is run - if zuul-cloner is used.

We run jobs on long lived workers and also on developers machines, let's
cleanup afterwards.

Install into a temporary directory that can be removed with "git clean"
afterwards. Add this directory to ignore list of pep8.

Simplify with using pushd/popd which are bash features, so change
she-bang.

Also, remove setup of ZUUL_BRANCH, this is not needed with current zuul
anymore.

Change-Id: I54d400dea4efdf9ac9b580ff7f0a2bbcceb4df5e