Prevent LBaaSv2 agent from filling up logs.
The message concerning stats sock not available is now only printed
when the lbaas-listener is created (has an assigned VIP) or when in
debug level.

The needed stats sock is only present when the loadbalancer has a VIP
assigned/the lbaas-listener is created.

Unit tests also updated to match new code.

Change-Id: I37785c33422e09eba6bf6ff26598919945dcf061
Closes-Bug: #1569827