Reduce the amount of smoke tests from 109 to 36
Smoke testing is designed around a compromise in testing quality in
order to have results in a way that consumes less resources and time.

We want to be able to have a reasonable confidence that the service
works fairly well without having to run too many tests.

As such, let's keep what still amounts to a fairly good testing of
each component tagged as smoke.
More in-depth testing can still be achieved with the right pattern
matching but smoke testing will not be as lengthy and resource
consuming now.

Change-Id: Ic1ef8a2f7fd0dc5a19e649eeb67ad4960df73360