Don not create loadbalancer if no eligible lbaasv2 agent
If we create a loadbalancer when all lbaasv2 agent is down,
the server return error with 404 http code but the loadbalancer
has been created in the database. And the created loadbalancer
can nerver be ACTIVE after all lbaasv2 agent starting up because
it hasn't been hosted to a certain lbaasv2 agent. What is more,
this loadbalancer can not be deleted due to its PENDING_CREATE status.

We can set the status of the creating loadbalaner to ERROR if there
is no eligible lbaasv2 agent, but I think it's better that we
don't create loadbalancers if there is no eligible lbaasv2 agent.

Change-Id: Ib45d73de323ba736bcef0fbf39ab3f97f1faa21c
Closes-Bug: #1652702