Fix KeyError exception in ChanceScheduler.schedule
The return value of function "get_agent_hosting_loadbalancer" is either
None or a dict. Most of the time, it looks good because it usually
returns None. But when multi neutron-server rescheduled the same
loadbalancer from a dead agent almost simultaneously, it might return a
dict, which would cause KeyError since the dict format is such like
{'agent': {'id': xxx, ...}}.

Change-Id: Iac0f99590146ef870a979a6b4ac28b69e9e186e4
Closes-Bug: #1668513