Fix the wrong device_owner when recreate LBaas listener
Currently, lbaas agent refresh loadbalancer based on the number of active
listener, and will update the port with device_own='' if delete the last
listener in loadbalancer.

So the device_owner of the vip port of exist loadbalancer will never
update to the correct one.

This patch add port device_owner field to PUT to guarantee the field is
'neutron:LOADBALANCERV2'.

Change-Id: I2de5800f4a6b06f98174ef42390f0e2c91bf0cfd
Closes-Bug:#1692995