Fix a namespace collision issue.
Both the neutron-lbaas and neutron tests have a new_update_request method.
The neutron version is brought in via a mixin where the neutron-lbaas
verision is in the base unit test class.

This is a quick and dirty fix by renaming the neutron-lbaas version to
get the gates functioning again.

Change-Id: Ie89fda13ce5917aa9e0eb50125b29b77a5d787c9