Fix tempest scenario tests waiters
Cleanups for resources such as listeners, pools and health monitors are
currently using a waiter named _wait_for_load_balancer_status to get an
indication that the cleanup process ended successfully.

Because of the way that waiter is currently used in those cleanups,
it's basically just checking that the load balancer state is:
provisioning_status='ACTIVE'
operating_status='ONLINE'

This is not a good measure for the success of a cleanup for the
aforementioned resources since the deletion process of those simply has
no effect on either loadbalancer provisioning_status or
operating_status. Thus, this is not a good indicator.

This creates issues on high load test deployments.
For example: test_listener_basic is a scenario that deletes a listener
and verifies the absence of the load balancing functionality via
_check_load_balancing_after_deleting_resources.
In case the listener deletion is still undergoing (and currently it
does not have a good indication for that), the load balancer will
accept the traffic and fail the test.

The suggestion solution here is not to use _wait_for_load_balancer_status
as an indicator, but rather add waiters to check specifically if a given
resource got deleted or not.

Change-Id: Ib2e4c60e2db28f1345665c2f5859e8f5fde75bc0