Update pool member operating status for haproxy
Collection of stats is a period task for which the HAProxy driver also
returns member statuses, including 'status' field which maps to
MemberV2.operating_status. Updating such field in neutron-lbaas DB
ensures the stored operational status is up-to-date. Subsequently, the
operating status of pools, listeners and load balancers can also be
correctly reported to users upon querying in for the load balancer
status tree API.

This fixes cases where a pool member becomes offline in the dataplane
but is not observed in neutron-lbaas. If a member turns back online
again, its operating status will be updated too.

Partial-Bug: #1548774

Change-Id: Ief872b0463002b4e339f8eab71b37a4225d461cc