Fix "unable to delete subnet" in API tests
We have been seeing occasional "unable to delete subnet, One or more ports
have an IP allocation from this subnet." errors in the
neutron-lbaasv2-dsvm-api tests. This tends to happen on slower test hosts.
My suspicion is that since the port deletion is asynchronous, we are trying
to delete the subnet before neutron finishes the port deletion.
The neutron logs [1] appear to back this theory.

This patch adds a wait for the port to finish deletion in neutron before
moving on to delete the subnet.

[1] http://logs.openstack.org/68/539068/2/check/neutron-lbaasv2-dsvm-api/ \
    6fba02b/logs/screen-q-svc.txt.gz#_Jan_30_21_33_58_634644

Change-Id: I3baea0319114c9c5a2f8839d4a6b90f9022fefe7