Log failed cert delete, and raise TLSContainerInvalid instead
It's possible that neutron can get a certificate, but not delete it. This
can cause the cert validation to raise a HTTP error instead of the
actual validation error.

This patch changes this by catching the cert_mgr error and logging it,
allowing the original error to raise properly.

Change-Id: Icd8fd77a157481e3ce19ea71749bb6e0e3940967
(cherry picked from commit 5b58a20906a13cd141d39d4684aa8c830fa174da)