Fix proxy extension for neutron RBAC
Neutron is looking for 'tenant_id' in the response from the proxy plugin.
This patch makes sure the 'project_id' returned from octavia is copied over
to the 'tenant_id' in the response to neutron or assigns the one from the
context.

This patch also adapts Octavia's 403 Forbidden errors to neutron's 404 [1].
[1] https://docs.openstack.org/neutron/pike/contributor/internals/ \
    policy.html#request-authorization

Neutron is also returning a 404 instead of a status "DELETED". This
adjusts that.

Neutron/tempest will ask for the tenant_id field - replace this with
project_id.

Handle 403 Forbidden

Fixes bug to not pass on empty string to Octavia

Story: 2002602
Task: 22220

Change-Id: I840b5e2e8a95a4b3e166db9f9aa44590441aa7b6