Cap haproxy log level severity
With current haproxy config, emerg logs are sent to syslog when there
are no backends available. Emerg logs on most syslog setups are sent
via wall to all users.

This is rarely a wanted behavior. And it might happen quite often
at the configuration phase.

Cap logs to `alert` level. This converts `emerg` logs to `alert`,
which by default are logged to files, and prevents wall-like messages.

This fix is also suggested by haproxy docs:

https://cbonte.github.io/haproxy-dconv/1.7/configuration.html#4.2-log

Change-Id: I4c5ede6d3ef2479fb123dbb6706a7d5090b6e5ec
Closes-Bug: 1362213