Accommodate for neutron code that expires obsolete relationships
Neutron is going to introduce a change that will make sqlalchemy to
automatically expire models reached through obsolete / invalid foreign keys.
Because of that, we need to recreate policy models after they are popped up
from the policies' list because old models immediately become expired.

Also, this patch adds a new backref to LoadBalancer model that reverse-links it
to provider model. This is also needed to accommodate for the neutron change
that assumes a relationship is always to have a reverse property linking back
to parent / child model.

Change-Id: If4b28110f460f6ac77ace1bbb02967ea986d4cab
Needed-By: I0d65d19204da8ce30addfa5faff68544534b7853