Updated "create_pool" method in plugin
Once the pool object is created in the DB and the listeners
are updated with the pool information, the updated listeners
are retrieved and populated into the pool object.

The DB method to create the pool now flushes the context to
retrieve an updated pool DB object at the end of the method,
instead of retrieving again the same object from the DB.

This is needed to make solve the Queens backport of
I32328c5206b9cd6fb8d8764c079f22b6ea8bfa9e.

Story: 2001962

Change-Id: Icf43952f6e5c5004a5c4d8d541b6efcb5d3ec50c